Cmd refactor
