--- code/sys.driver/driver-vitepress/src/-cmd/common.ts
@@ -1,7 +1,0 @@
-export * from '../common.ts';
-
-export const DEFAULTS = {
-  cmd: 'dev',
-  inDir: './',
-  open: undefined,
-} as const;
