+++ code/sys.driver/driver-vitepress/src/m.Cmd/common.ts
@@ -1,1 +1,7 @@
 export * from '../common.ts';
+
+export const DEFAULTS = {
+  cmd: 'dev',
+  inDir: './',
+  open: undefined, // NB: use default.
+} as const;
