+++ code/sys.driver/driver-vitepress/src/m.VitePress/mod.ts
@@ -1,5 +1,6 @@
 import type { t } from './common.ts';
 
+import { Cmd } from '../m.Cmd/mod.ts';
 import { Env } from '../m.Env/mod.ts';
 import { build } from './u.cmd.build.ts';
 import { dev } from './u.cmd.dev.ts';
@@ -9,6 +10,7 @@ import { dev } from './u.cmd.dev.ts';
  */
 export const VitePress: t.VitePressLib = {
   Env,
+  Cmd,
   dev,
   build,
 };
