+++ code/sys.driver/driver-vitepress/src/types.ts
@@ -2,6 +2,6 @@
  * @module
  * Module types.
  */
-export type * from './-cmd/t.ts';
+export type * from './m.Cmd/t.ts';
 export type * from './m.Env/t.ts';
 export type * from './m.VitePress/t.ts';
